feat: 폴더 페이지 루트 경로, id 경로 연동, 폴더 선택 버튼 연동
